requirements.txt
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.1
django-cors-headers==2.1.0
django-extensions==1.9.9
django-heroku==0.3.1
djangorestframework==3.7.7
gunicorn==19.9.0
pytz==2018.7
requests==2.21.0
six==1.12.0
typing==3.6.6
urllib3==1.24.1
whitenoise==4.1.2
